Move token-specific functions into `Parser.ts`
