+++ src/files/BrsFile.ts
@@ -117,18 +117,6 @@ export class BrsFile {
 
     private workspaceSymbols: SymbolInformation[];
 
-    /**
-     * Get the token at the specified position
-     * @param position
-     */
-    public getTokenAt(position: Position) {
-        for (let token of this.parser.tokens) {
-            if (util.rangeContains(token.range, position)) {
-                return token;
-            }
-        }
-    }
-
     public get parser() {
         if (!this._parser) {
             //remove the typedef file (if it exists)
@@ -545,7 +533,7 @@ export class BrsFile {
         }
 
         //if cursor is within a comment, disable completions
-        let currentToken = this.getTokenAt(position);
+        let currentToken = this.parser.getTokenAt(position);
         const tokenKind = currentToken?.kind;
         if (tokenKind === TokenKind.Comment) {
             return [];
@@ -587,7 +575,7 @@ export class BrsFile {
         let functionExpression = this.getFunctionExpressionAtPosition(position);
         if (!functionExpression) {
             //we aren't in any function scope, so return the keyword completions and namespaces
-            if (this.getTokenBefore(currentToken, TokenKind.New)) {
+            if (this.parser.getTokenBefore(currentToken, TokenKind.New)) {
                 // there's a new keyword, so only class types are viable here
                 return [...this.getGlobalClassStatementCompletions(currentToken, this.parseMode)];
             } else {
@@ -596,7 +584,7 @@ export class BrsFile {
         }
 
         const classNameCompletions = this.getGlobalClassStatementCompletions(currentToken, this.parseMode);
-        const newToken = this.getTokenBefore(currentToken, TokenKind.New);
+        const newToken = this.parser.getTokenBefore(currentToken, TokenKind.New);
         if (newToken) {
             //we are after a new keyword; so we can only be namespaces or classes at this point
             result.push(...classNameCompletions);
@@ -604,11 +592,11 @@ export class BrsFile {
             return result;
         }
 
-        if (this.tokenFollows(currentToken, TokenKind.Goto)) {
+        if (this.parser.tokenFollows(currentToken, TokenKind.Goto)) {
             return this.getLabelCompletion(functionExpression);
         }
 
-        if (this.isPositionNextToTokenKind(position, TokenKind.Dot)) {
+        if (this.parser.isPositionNextToTokenKind(position, TokenKind.Dot)) {
             if (namespaceCompletions.length > 0) {
                 //if we matched a namespace, after a dot, it can't be anything else but something from our namespace completions
                 return namespaceCompletions;
@@ -712,9 +700,9 @@ export class BrsFile {
     }
 
     public getClassFromMReference(position: Position, currentToken: Token, functionExpression: FunctionExpression): FileLink<ClassStatement> | undefined {
-        let previousToken = this.getPreviousToken(currentToken);
+        let previousToken = this.parser.getPreviousToken(currentToken);
         if (previousToken?.kind === TokenKind.Dot) {
-            previousToken = this.getPreviousToken(previousToken);
+            previousToken = this.parser.getPreviousToken(previousToken);
         }
         if (previousToken?.kind === TokenKind.Identifier && previousToken?.text.toLowerCase() === 'm' && isClassMethodStatement(functionExpression.functionStatement)) {
             return { item: this.parser.references.classStatements.find((cs) => util.rangeContains(cs.range, position)), file: this };
@@ -761,7 +749,7 @@ export class BrsFile {
         //remove any trailing identifer and then any trailing dot, to give us the
         //name of its immediate parent namespace
         let closestParentNamespaceName = completionName.replace(/\.([a-z0-9_]*)?$/gi, '');
-        let newToken = this.getTokenBefore(currentToken, TokenKind.New);
+        let newToken = this.parser.getTokenBefore(currentToken, TokenKind.New);
 
         let namespaceLookup = scope.namespaceLookup;
         let result = new Map<string, CompletionItem>();
@@ -875,64 +863,6 @@ export class BrsFile {
         }
     }
 
-    public isPositionNextToTokenKind(position: Position, tokenKind: TokenKind) {
-        const closestToken = this.getClosestToken(position);
-        const previousToken = this.getPreviousToken(closestToken);
-        const previousTokenKind = previousToken?.kind;
-        //next to matched token
-        if (!closestToken || closestToken.kind === TokenKind.Eof) {
-            return false;
-        } else if (closestToken.kind === tokenKind) {
-            return true;
-        } else if (closestToken.kind === TokenKind.Newline || previousTokenKind === TokenKind.Newline) {
-            return false;
-            //next to an identifier, which is next to token kind
-        } else if (closestToken.kind === TokenKind.Identifier && previousTokenKind === tokenKind) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    private getTokenBefore(currentToken: Token, tokenKind: TokenKind): Token {
-        const index = this.parser.tokens.indexOf(currentToken);
-        for (let i = index - 1; i >= 0; i--) {
-            currentToken = this.parser.tokens[i];
-            if (currentToken.kind === TokenKind.Newline) {
-                break;
-            } else if (currentToken.kind === tokenKind) {
-                return currentToken;
-            }
-        }
-        return undefined;
-    }
-
-    private tokenFollows(currentToken: Token, tokenKind: TokenKind): boolean {
-        const index = this.parser.tokens.indexOf(currentToken);
-        if (index > 0) {
-            return this.parser.tokens[index - 1].kind === tokenKind;
-        }
-        return false;
-    }
-
-    public getTokensUntil(currentToken: Token, tokenKind: TokenKind, direction: -1 | 1 = -1) {
-        let tokens = [];
-        for (let i = this.parser.tokens.indexOf(currentToken); direction === -1 ? i >= 0 : i === this.parser.tokens.length; i += direction) {
-            currentToken = this.parser.tokens[i];
-            if (currentToken.kind === TokenKind.Newline || currentToken.kind === tokenKind) {
-                break;
-            }
-            tokens.push(currentToken);
-        }
-        return tokens;
-    }
-
-    public getPreviousToken(token: Token) {
-        const parser = this.parser;
-        let idx = parser.tokens.indexOf(token);
-        return parser.tokens[idx - 1];
-    }
-
     /**
      * Find the first scope that has a namespace with this name.
      * Returns false if no namespace was found with that name
@@ -974,29 +904,6 @@ export class BrsFile {
             }
         }
         return false;
-    }
-
-    /**
-     * Get the token closest to the position. if no token is found, the previous token is returned
-     * @param position
-     * @param tokens
-     */
-    public getClosestToken(position: Position) {
-        let tokens = this.parser.tokens;
-        for (let i = 0; i < tokens.length; i++) {
-            let token = tokens[i];
-            if (util.rangeContains(token.range, position)) {
-                return token;
-            }
-            //if the position less than this token range, then this position touches no token,
-            if (util.positionIsGreaterThanRange(position, token.range) === false) {
-                let t = tokens[i - 1];
-                //return the token or the first token
-                return t ? t : tokens[0];
-            }
-        }
-        //return the last token
-        return tokens[tokens.length - 1];
     }
 
     /**
@@ -1121,7 +1028,7 @@ export class BrsFile {
         let results: Location[] = [];
 
         //get the token at the position
-        const token = this.getTokenAt(position);
+        const token = this.parser.getTokenAt(position);
 
         // While certain other tokens are allowed as local variables (AllowedLocalIdentifiers: https://github.com/rokucommunity/brighterscript/blob/master/src/lexer/TokenKind.ts#L418), these are converted by the parser to TokenKind.Identifier by the time we retrieve the token using getTokenAt
         let definitionTokenTypes = [
@@ -1136,7 +1043,7 @@ export class BrsFile {
 
         let textToSearchFor = token.text.toLowerCase();
 
-        const previousToken = this.getTokenAt({ line: token.range.start.line, character: token.range.start.character });
+        const previousToken = this.parser.getTokenAt({ line: token.range.start.line, character: token.range.start.character });
 
         if (previousToken?.kind === TokenKind.Callfunc) {
             for (const scope of this.program.getScopes()) {
@@ -1149,7 +1056,7 @@ export class BrsFile {
             return results;
         }
 
-        let classToken = this.getTokenBefore(token, TokenKind.Class);
+        let classToken = this.parser.getTokenBefore(token, TokenKind.Class);
         if (classToken) {
             let cs = this.parser.references.classStatements.find((cs) => cs.classKeyword.range === classToken.range);
             if (cs?.parentClassName) {
@@ -1183,7 +1090,7 @@ export class BrsFile {
                 results.push(Location.create(uri, symbol.range));
             }
         }
-        if (this.tokenFollows(token, TokenKind.Goto)) {
+        if (this.parser.tokenFollows(token, TokenKind.Goto)) {
             for (const label of func.labelStatements) {
                 if (label.tokens.identifier.text.toLocaleLowerCase() === textToSearchFor) {
                     const uri = util.pathToUri(this.srcPath);
@@ -1250,7 +1157,7 @@ export class BrsFile {
 
     public getHover(position: Position): Hover {
         //get the token at the position
-        let token = this.getTokenAt(position);
+        let token = this.parser.getTokenAt(position);
 
         let hoverTokenTypes = [
             TokenKind.Identifier,
@@ -1361,10 +1268,10 @@ export class BrsFile {
         const funcStartPosition = func.range.start;
 
         // Get function comments in reverse order
-        let currentToken = this.getTokenAt(funcStartPosition);
+        let currentToken = this.parser.getTokenAt(funcStartPosition);
         let functionComments = [] as string[];
         while (currentToken) {
-            currentToken = this.getPreviousToken(currentToken);
+            currentToken = this.parser.getPreviousToken(currentToken);
 
             if (!currentToken) {
                 break;
@@ -1450,7 +1357,7 @@ export class BrsFile {
 
     public getReferences(position: Position) {
 
-        const callSiteToken = this.getTokenAt(position);
+        const callSiteToken = this.parser.getTokenAt(position);
 
         let locations = [] as Location[];
